refactor: project settings member
