+++ web/components/project/index.ts
@@ -14,3 +14,8 @@ export * from "./priority-select";
 export * from "./sidebar-list-item";
 export * from "./sidebar-list";
 export * from "./integration-card";
+export * from "./member-list";
+export * from "./member-list-item";
+export * from "./project-settings-member-defaults";
+export * from "./send-project-invitation-modal";
+export * from "./confirm-project-member-remove";
